Split meld-system into detect/feedback/composition modules
Barrel re-export preserves all existing import paths. Updated
implementation checklist for Phase 3 completion.
